Basic hazards created
Created abstract class "Hazard" which will function as a parent for the various hazards in the game. Created several hazard instances as well. Now we simply need better implementation of the hazards through procedural generation, check / account for the fail state, and manage scoring in order to have a playable product.